+++ packages/client/src/play/index.js
@@ -1,6 +1,7 @@
 export { createLivePlayClient } from './live.js'
 export * from './chrome.js'
 export * from './create.js'
+export * from './fork.js'
 export * from './occupancy.js'
 export * from './sidebar-model.js'
 export { PlayWorkspaceBrowser } from './sidebar.js'